Changed name of pl_stdio_t to gs_lib_ctx_t.
Changed stream code to always keep a valid gs_memory_t ptr
and keep a seperate allocated on the heap flag.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1977 06663e23-700e-0410-b217-a244a6096597

+++ gs/src/gxcindex.h
@@ -43,7 +43,7 @@
  */
 
 #ifdef TEST_CINDEX_STRUCT
-
+#error // stefan
 /* Define the type for device color index (pixel value) data. */
 typedef struct { ulong value[2]; } gx_color_index_data;
 
@@ -51,9 +51,14 @@ typedef struct { ulong value[2]; } gx_color_index_data;
 
 /* Define the type for device color index (pixel value) data. */
 #ifdef GX_COLOR_INDEX_TYPE
+
 typedef GX_COLOR_INDEX_TYPE gx_color_index_data;
 #else
-typedef ulong gx_color_index_data;
+
+// stefan toplevel makefile needs to define
+// -DGX_COLOR_INDEX_TYPE='unsigned long long'
+
+typedef unsigned long long gx_color_index_data;
 #endif
 
 #endif /* (!)TEST_CINDEX_STRUCT */
